Move user avatar info to navigation drawer
