+++ App/src/main/java/cc/softwarefactory/lokki/android/activities/SettingsActivity.java
@@ -1,0 +1,16 @@
+package cc.softwarefactory.lokki.android.activities;
+
+import android.os.Bundle;
+import android.support.v7.app.ActionBarActivity;
+
+import cc.softwarefactory.lokki.android.R;
+
+public class SettingsActivity extends ActionBarActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.fragment_drawer_header);
+    }
+}
