--- App/src/main/res/layout/activity_settings.xml
@@ -1,74 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:layout_gravity="center"
-    android:background="@android:color/white"
-    android:orientation="vertical"
-    android:divider="?android:listDivider"
-    android:padding="0dp"
-    android:showDividers="middle">
-
-    <RelativeLayout
-        android:layout_gravity="center"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:context=".AboutActivity"
-        android:padding="16dp">
-
-    <com.makeramen.RoundedImageView xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/avatar"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="10dp"
-        android:scaleType="centerCrop"
-        android:src="@drawable/default_avatar"
-        app:border_color="#215fcc"
-        app:border_width="0dip"
-        app:corner_radius="100dip"
-        app:round_background="true" />
-
-    <TextView
-        android:id="@+id/user_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/avatar"
-        android:layout_centerHorizontal="true"
-        android:padding="10dip"
-        android:text=""
-        android:textColor="@android:color/black"
-        android:textSize="18sp" />
-
-    <TextView
-        android:id="@+id/lokki_id_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/user_name"
-        android:layout_centerHorizontal="true"
-        android:padding="0dip"
-        android:text=""
-        android:textColor="@android:color/black"
-        android:textSize="12sp" />
-
-    <TextView
-        android:id="@+id/your_friends"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/lokki_id_text"
-        android:layout_centerHorizontal="true"
-        android:padding="0dip"
-        android:text="@string/your_friends"
-        android:textColor="@android:color/darker_gray"
-        android:textSize="12sp"
-        android:gravity="center_horizontal" />
-    </RelativeLayout>
-
-    <fragment
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="0dp"
-        android:name="cc.softwarefactory.lokki.android.PreferencesFragment"
-        android:id="@+id/fragment_settings"/>
-
-</LinearLayout>
