+++ App/src/main/res/layout/fragment_drawer_header.xml
@@ -1,0 +1,42 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:layout_gravity="center"
+    android:background="@android:color/white"
+    android:orientation="vertical"
+    android:divider="?android:listDivider"
+    android:padding="0dp"
+    android:showDividers="middle">
+
+    <RelativeLayout
+        android:layout_gravity="left"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        tools:context=".AboutActivity"
+        android:padding="16dp">
+
+    <com.makeramen.RoundedImageView xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/avatar"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:scaleType="centerCrop"
+        android:src="@drawable/default_avatar"
+        app:border_color="#215fcc"
+        app:border_width="0dip"
+        app:corner_radius="100dip"
+        app:round_background="true" />
+
+    <TextView
+        android:id="@+id/user_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/avatar"
+        android:padding="10dip"
+        android:text=""
+        android:textColor="@android:color/black"
+        android:textSize="14sp" />
+
+    </RelativeLayout>
+
+</LinearLayout>
